Create if statement that checks the existence of the json file, assign contents to variable and print the contents to the screen. Move the code that writes the users' favorite number to a json file into the else statement.

+++ chapter-10/10-11_Favorite_Number.py
@@ -5,9 +5,15 @@ import json
 
 path = Path('user_favorite_number.json')
 
-fav_number = input("Tell me your favorite number and I'll remember it. ")
-contents = json.dumps(fav_number)
-path.write_text(contents)
+if path.exists():
+    contents = path.read_text()
+    fav_number = json.loads(contents)
+    print(f"I remember your favorite number! It's {fav_number}! ")
 
-print(
-    f"Your number is {fav_number}. I'll remember that the next time we meet.")
+else:
+    fav_number = input("Tell me your favorite number and I'll remember it. ")
+    contents = json.dumps(fav_number)
+    path.write_text(contents)
+
+    print(
+        f"Your number is {fav_number}. I'll remember that the next time we meet.")
